initial server database management

+++ WorldBuilder/Lib/Settings/SettingsMetaData.cs
@@ -17,7 +17,7 @@ namespace WorldBuilder.Lib.Settings {
         public SettingRangeAttribute? Range { get; }
         public SettingPathAttribute? Path { get; }
         public string? Format { get; }
-        public int Order { get; }
+        public double Order { get; }
         public bool IsHidden { get; }
 
         [UnconditionalSuppressMessage("Trimming", "IL2075:Selectively keep fields for settings", Justification = "Settings classes are preserved by SourceGenerationContext")]
@@ -63,7 +63,7 @@ namespace WorldBuilder.Lib.Settings {
         public Type Type { get; }
         public string Name { get; }
         public string? ParentCategory { get; }
-        public int Order { get; }
+        public double Order { get; }
         public List<SettingPropertyMetadata> Properties { get; }
         public List<SettingCategoryMetadata> SubCategories { get; }
